feat: Add react 18 support
