+++ examples/react-18/.umirc.ts
@@ -1,0 +1,13 @@
+import { defineConfig } from 'umi';
+
+export default defineConfig({
+  nodeModulesTransform: {
+    type: 'none',
+  },
+  routes: [
+    {
+      path: '/',
+      component: '@/pages/index',
+    },
+  ],
+});
