+++ packages/renderer-react/.fatherrc.ts
@@ -1,4 +1,10 @@
 export default {
+  entry: [
+    // For react <= 17
+    './src/index.ts',
+    // For react 18
+    './src/index18.ts',
+  ],
   target: 'browser',
   cjs: { type: 'rollup', lazy: false },
   esm: { type: 'rollup' },
