+++ packages/renderer-react/src/index18.ts
@@ -1,0 +1,2 @@
+export * from './commonExports';
+export { default as renderClient } from './renderClient/renderClient-18';
