+++ packages/renderer-react/src/renderClient/renderClient.tsx
@@ -1,69 +1,8 @@
-import { ApplyPluginsType, Plugin, Router } from '@umijs/runtime';
-import React, { useEffect } from 'react';
 import { hydrate, render } from 'react-dom';
 import { matchRoutes, RouteConfig } from 'react-router-config';
 import { IRoute } from '..';
-import renderRoutes from '../renderRoutes/renderRoutes';
-
-interface IRouterComponentProps {
-  routes: IRoute[];
-  plugin: Plugin;
-  history: any;
-  ssrProps?: object;
-  defaultTitle?: string;
-  dynamicImport?: boolean;
-  isServer?: boolean;
-}
-
-interface IOpts extends IRouterComponentProps {
-  rootElement?: string | HTMLElement;
-  callback?: () => void;
-}
-
-function RouterComponent(props: IRouterComponentProps) {
-  const { history, ...renderRoutesProps } = props;
-
-  useEffect(() => {
-    // first time using window.g_initialProps
-    // switch route fetching data, if exact route reset window.getInitialProps
-    if ((window as any).g_useSSR) {
-      (window as any).g_initialProps = null;
-    }
-    function routeChangeHandler(location: any, action?: string) {
-      const matchedRoutes = matchRoutes(
-        props.routes as RouteConfig[],
-        location.pathname,
-      );
-
-      // Set title
-      if (
-        typeof document !== 'undefined' &&
-        renderRoutesProps.defaultTitle !== undefined
-      ) {
-        document.title =
-          (matchedRoutes.length &&
-            // @ts-ignore
-            matchedRoutes[matchedRoutes.length - 1].route.title) ||
-          renderRoutesProps.defaultTitle ||
-          '';
-      }
-      props.plugin.applyPlugins({
-        key: 'onRouteChange',
-        type: ApplyPluginsType.event,
-        args: {
-          routes: props.routes,
-          matchedRoutes,
-          location,
-          action,
-        },
-      });
-    }
-    routeChangeHandler(history.location, 'POP');
-    return history.listen(routeChangeHandler);
-  }, [history]);
-
-  return <Router history={history}>{renderRoutes(renderRoutesProps)}</Router>;
-}
+import getRootContainer from './getRootContainer';
+import type { IOpts } from './types';
 
 /**
  * preload for SSR in dynamicImport
@@ -91,24 +30,7 @@ export async function preloadComponent(
 }
 
 export default function renderClient(opts: IOpts) {
-  const rootContainer = opts.plugin.applyPlugins({
-    type: ApplyPluginsType.modify,
-    key: 'rootContainer',
-    initialValue: (
-      <RouterComponent
-        history={opts.history}
-        routes={opts.routes}
-        plugin={opts.plugin}
-        ssrProps={opts.ssrProps}
-        defaultTitle={opts.defaultTitle}
-      />
-    ),
-    args: {
-      history: opts.history,
-      routes: opts.routes,
-      plugin: opts.plugin,
-    },
-  });
+  const rootContainer = getRootContainer(opts);
 
   if (opts.rootElement) {
     const rootElement =
